requirements.txt
streamlit==1.31.0
google-generativeai==0.8.0
newspaper3k==0.2.8
duckduckgo_search==4.1.1
lxml_html_clean==0.1.1
lxml==5.1.0
